Bring in Knockout, underscore, basket.
Knockout for client side templates and MVC-ish updating.  Underscore
because I like it and use it a lot.  Basket for ease of offline
development (and if others would be working on this codebase it helps to
get them up to speed if all of the dependencies are already a part of the
basic checkout.)