Basic AJAX and DOM update.
Just making sure that the guard() function works.  I wouldn't want to
trust Flask's auth/session/cookie stuff without double checking how it
works, but it seems fine for now.  (I used wget to try hitting the
'/todos' endpoint without being logged in and it is rejected.)

So, now we have AJAX that respects the logged in user and we can get down
to business.