Inlining the JS libs for a bit of performance.
Basically my laptop is slow enough that serving the JS as separate files
and incurring the additional connections is noticable (1 second+ !) so
let's just tuck it all into the HTML directly.  It's not that much by
modern standards and it only has to load once the first time you hit '/'
and then the rest of the app works by AJAX.

This might be silly if we were loading a ton of stuff, but the three libs
I'm using total under 87K.  My gravatar image is more than that. ;-)

In other news the developers of gedit are clowns who can't program their
way out of a paper bag.  Long lines make gedit cry and cut itself. O_o