Apparently jinja hates knockout.
You can't just include knockout in the template because it uses {{}} in
there and that confuses jinja.  So crude hack to stuff it into a template
variable and interpolate it at serving-time.  This gives a sad, but on
the bright side including all the JS in the HTML page brought load time
down to ~40ms (on my crappy slow 2-core 1.6Ghz laptop!) so over all it's
a win even with this monkey-crap.