Basic display of todos.
New todos are appended to the list, while updates replace themselves.

(Later it might make sense to just always append if we're going to
trigger a re-sort after modifying the list anyway...)

Also, I have a 'sync' attribute that is meant to indicate that a given
todo on the client-side is actually in the db.  When the user creates a
new todo it should be stuck in the list with sync=false and then if the
AJAX request to put it in the db completes we can flip it to true.  Same
for making updates to existing todos.  Go ahead and update the client but
don't set sync=true until the request returns sucessfully.

Once that is in place I'm going to set up a little watcher to periodially
check for un-sync-ed todos in the list and automatically try to sync them
with the server; showing a little "uh-oh" warning if we can't reach
the server for some reason.

As a bonus, knockout makes it easy to have a little icon or style effect
to indicate the sync'd-ness of the todos.