Due dates, fin deletion.
Okay, so, yes, seconds-since-the-Epoch plus timezone isn't an ideal
display format. I get that.  The moment.js lib will allow for nicer
formatting once I readthedocs, the trick is to trigger the updates to
happen just after knockout renders teh html, you can't just add a binding
via zepto/jQuery at start time because the html tiles we want to modify
are being created later in the app's lifecycle.  I'm pretty sure KO
allows for post-processing after updates.  It isn't easy to do this when
the docs are all on teh internet and your connection suks.  But the main
bits are there now: date and tz are making the round trip into the db and
back to the UI and we're *sure* they're not getting gummed up anywhere.

We make pretty later.