Can modify entries.
Tiles now have an "Edit" button that works by populating the form with
the data from the tile's todo, and it modifies the behaviour of the
next_id() function to return that ID instead of calculating the next ID.

When you post the data it modifies the record in the db and returns the
jsonish version which then triggers the update on the client.

It would be easy to add a "Cancel" button that only appears if the
next_id() funtion is set to do this (using knockout to make the button
visible when VM.nid is NaN,)  It would just set nid=Nan.

But anyway, editing FTW.